+++ src/analog/m2kanalogout.cpp
@@ -37,22 +37,22 @@ void M2kAnalogOut::init()
 	m_pimpl->init();
 }
 
-std::vector<double> M2kAnalogOut::getOversamplingRatio()
+std::vector<int> M2kAnalogOut::getOversamplingRatio()
 {
 	return m_pimpl->getOversamplingRatio();
 }
 
-double M2kAnalogOut::getOversamplingRatio(unsigned int chn_idx)
+int M2kAnalogOut::getOversamplingRatio(unsigned int chn_idx)
 {
 	return m_pimpl->getOversamplingRatio(chn_idx);
 }
 
-std::vector<double> M2kAnalogOut::setOversamplingRatio(std::vector<double> oversampling_ratio)
+std::vector<int> M2kAnalogOut::setOversamplingRatio(std::vector<int> oversampling_ratio)
 {
 	return m_pimpl->setOversamplingRatio(oversampling_ratio);
 }
 
-double M2kAnalogOut::setOversamplingRatio(unsigned int chn_idx, double oversampling_ratio)
+int M2kAnalogOut::setOversamplingRatio(unsigned int chn_idx, int oversampling_ratio)
 {
 	return m_pimpl->setOversamplingRatio(chn_idx, oversampling_ratio);
 }
